Added diplay of player params in-game, added API for RL model

+++ Driving-Game/Code/PlayerData.cs
@@ -3,6 +3,6 @@ public class PlayerData
     public int Rotation { get; set; }
     public int Slope { get; set; }
     public int DistToGround { get; set; }
-    public int AngularVelocity { get; set; }
+    public double AngularVelocity { get; set; }
     public bool IsTouchingGround { get; set; }
 }
